Add files via upload
The Data Encryption Standard (DES) is a symmetric-key algorithm for encryption and decryption of data. It was widely used in the 1970s and 1980s for secure communication and data protection. The algorithm uses a 56-bit key to encrypt and decrypt data in 64-bit blocks.

The encryption process involves a series of permutations and substitutions of the input data and key. The input data is divided into 64-bit blocks and is subjected to an initial permutation. The key is then used to generate 16 round keys, each of which is used in a series of substitution and permutation operations on the data. Finally, a final permutation is applied to the output to generate the encrypted data.

The decryption process is essentially the same as the encryption process, but with the round keys used in reverse order. The algorithm is considered secure, but due to its relatively short key length, it is vulnerable to brute force attacks. Therefore, it has been largely replaced by more secure algorithms such as AES.